// web/packages/web-client/src/graph/generated/models/Identity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Libre Graph API
 * Libre Graph is a free API for cloud collaboration inspired by the MS Graph API.
 *
 * The version of the OpenAPI document: v1.0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Identity
 */
export interface Identity {
    /**
     * The identity's display name. Note that this may not always be available or up to date. For example, if a user changes their display name, the API may show the new value in a future response, but the items associated with the user won't show up as having changed when using delta.
     */
    displayName: string;
    /**
     * Unique identifier for the identity.
     */
    id?: string;
    /**
     * The type of the identity. This can be either "Member" for regular user, "Guest" for guest users or "Federated" for users imported from a federated instance. Can be used by clients to indicate the type of user. For more details, clients should look up and cache the user at the /users endpoint.
     */
    atLibreGraphUserType?: string;
}

/**
 * Check if a given object implements the Identity interface.
 */
export function instanceOfIdentity(value: object): value is Identity {
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    return true;
}

export function IdentityFromJSON(json: any): Identity {
    return IdentityFromJSONTyped(json, false);
}

export function IdentityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Identity {
    if (json == null) {
        return json;
    }
    return {
        
        'displayName': json['displayName'],
        'id': json['id'] == null ? undefined : json['id'],
        'atLibreGraphUserType': json['@libre.graph.userType'] == null ? undefined : json['@libre.graph.userType'],
    };
}

export function IdentityToJSON(json: any): Identity {
    return IdentityToJSONTyped(json, false);
}

export function IdentityToJSONTyped(value?: Identity | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'displayName': value['displayName'],
        'id': value['id'],
        '@libre.graph.userType': value['atLibreGraphUserType'],
    };
}