// web/packages/web-client/src/graph/generated/models/UnifiedRoleDefinition.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Libre Graph API
 * Libre Graph is a free API for cloud collaboration inspired by the MS Graph API.
 *
 * The version of the OpenAPI document: v1.0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UnifiedRolePermission } from './UnifiedRolePermission';
import {
    UnifiedRolePermissionFromJSON,
    UnifiedRolePermissionFromJSONTyped,
    UnifiedRolePermissionToJSON,
    UnifiedRolePermissionToJSONTyped,
} from './UnifiedRolePermission';

/**
 * A role definition is a collection of permissions in libre graph listing the operations that can be performed
 * and the resources against which they can performed.
 * 
 * @export
 * @interface UnifiedRoleDefinition
 */
export interface UnifiedRoleDefinition {
    /**
     * The description for the unifiedRoleDefinition.
     */
    description?: string;
    /**
     * The display name for the unifiedRoleDefinition. Required. Supports $filter (`eq`, `in`).
     */
    displayName?: string;
    /**
     * The unique identifier for the role definition. Key, not nullable, Read-only. Inherited from entity. Supports $filter (`eq`, `in`).
     */
    id?: string;
    /**
     * List of permissions included in the role.
     */
    rolePermissions?: Array<UnifiedRolePermission>;
    /**
     * When presenting a list of roles the weight can be used to order them in a meaningful way.
     * Lower weight gets higher precedence. So content with lower weight will come first. If set,
     * weights should be non-zero, as 0 is interpreted as an unset weight.
     * 
     */
    atLibreGraphWeight?: number;
}

/**
 * Check if a given object implements the UnifiedRoleDefinition interface.
 */
export function instanceOfUnifiedRoleDefinition(value: object): value is UnifiedRoleDefinition {
    return true;
}

export function UnifiedRoleDefinitionFromJSON(json: any): UnifiedRoleDefinition {
    return UnifiedRoleDefinitionFromJSONTyped(json, false);
}

export function UnifiedRoleDefinitionFromJSONTyped(json: any, ignoreDiscriminator: boolean): UnifiedRoleDefinition {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'displayName': json['displayName'] == null ? undefined : json['displayName'],
        'id': json['id'] == null ? undefined : json['id'],
        'rolePermissions': json['rolePermissions'] == null ? undefined : ((json['rolePermissions'] as Array<any>).map(UnifiedRolePermissionFromJSON)),
        'atLibreGraphWeight': json['@libre.graph.weight'] == null ? undefined : json['@libre.graph.weight'],
    };
}

export function UnifiedRoleDefinitionToJSON(json: any): UnifiedRoleDefinition {
    return UnifiedRoleDefinitionToJSONTyped(json, false);
}

export function UnifiedRoleDefinitionToJSONTyped(value?: UnifiedRoleDefinition | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'description': value['description'],
        'displayName': value['displayName'],
        'id': value['id'],
        'rolePermissions': value['rolePermissions'] == null ? undefined : ((value['rolePermissions'] as Array<any>).map(UnifiedRolePermissionToJSON)),
        '@libre.graph.weight': value['atLibreGraphWeight'],
    };
}